requirements.txt
cycler==0.10.0
kiwisolver==1.2.0
llvmlite==0.33.0
matplotlib==3.3.0
numba==0.50.0
numpy==1.19.0
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
scipy==1.5.1
six==1.15.0